+++ .vscode/tasks.json
@@ -7,7 +7,8 @@
             "command": "g++",
             "args": [
                 "-g",                             // Enable debug symbols
-                "main.cpp",
+                "-Iinclude",
+                "src/*.cpp",
                 "-o",
                 "main",
                 "-lpdcurses"
