+++ include/enums.h
@@ -1,0 +1,15 @@
+#pragma once
+#ifndef ENUMS_H
+#define ENUMS_H
+
+enum MODE {
+    MOVING, INTERACTING
+};
+
+enum COLOR {
+    NORMAL = 1,
+    SELECTED_MOVING = 2,
+    SELECTED_INTERACTING = 3
+};
+
+#endif
