+++ src/main.cpp
@@ -1,276 +1,11 @@
-#include <curses.h>
-#include <iostream>
-#include <vector>
-#include <chrono>
-
-enum MODE {
-    MOVING, INTERACTING
-};
-
-enum COLOR {
-    NORMAL = 1,
-    SELECTED_MOVING = 2,
-    SELECTED_INTERACTING = 3
-};
-
-class RipplePrint;
-
-class WindowClass {
-    private:
-        bool has_border, isShowing, isSelectable;
-        int cur_w, cur_h, x, y, std_unit;
-        COLOR BorderColor;
-        WINDOW *win;
-        void (*content)(WindowClass * win); // Proper function pointer declaration
-        std::vector<RipplePrint*> Ripples;
-
-        std::vector<WindowClass> * Windows;
-
-    public:
-        WindowClass(std::vector<WindowClass> * _Windows, const int _std_unit, int height, int width, int _y, int _x,  void (*_content)(WindowClass * parent) = nullptr,  bool isSelectable = true, bool isShowing = true, bool has_border = true) : 
-            Windows(_Windows),
-            std_unit(_std_unit),
-            y(_y),
-            x(_x),
-            has_border(has_border),
-            isShowing(isShowing),
-            isSelectable(isSelectable),
-            content(_content) // Direct initialization
-        {
-            win = nullptr;
-            cur_h = height * std_unit;
-            cur_w = width * std_unit * 2;
-            y = _y * std_unit;
-            x = _x * std_unit * 2;
-
-            Windows->push_back(*this);
-        }
-
-        std::vector<RipplePrint*> getRipples();
-        void setBorderColor(COLOR color);
-        void AddRipple(RipplePrint* rip);
-        WINDOW * getWindow();
-        void InitWindow();
-        void Normalize(int x_offset, int y_offset);
-        void setSize(int height_stdunit, int width_stdunit);
-        void setPos (int x_stdunit, int y_stdunit);
-        int getRippleCount();
-        int getHeight();       
-        int getWidth();
-        int getX();
-        int getY();
-        bool IsShowing();
-        void IsShowing(bool set);
-        bool IsSelectable();
-        void IsSelectable (bool set) ;
-        void Update();
-};
+#include "WindowClass.h"
+#include "RipplePrint.h"
 
 void TimedErrorExit(std::string message, int countdownTime=3);
 void SetTerminalSize(std::vector<WindowClass>& Windows);
 
 std::vector<WindowClass> Windows;
 std::vector<std::vector<int>> WindowLayout;
-
-class RipplePrint 
-{
-    private:
-        std::string text, nowText;
-        int progress, x, y;
-        bool done, looping;
-        double delay, interval;
-        COLOR color;
-        std::chrono::_V2::system_clock::time_point startTime;
-        WindowClass * parent;
-        WINDOW * win;
-
-    public:
-        std::chrono::_V2::system_clock::time_point getStartTime() {
-            return startTime;
-        }
-
-        int getX() {
-            return x;
-        }
-
-        int getY() {
-            return y;
-        }
-
-        std::string getText() {
-            return text;
-        }
-
-        RipplePrint(WindowClass * _parent, int _y, int _x, std::string _text,double _interval = -1, double _delay=0.05, COLOR _color=COLOR::NORMAL) : 
-            parent(_parent),
-            text(_text),
-            x(_x),
-            y(_y),
-            delay(_delay),
-            color(_color),
-            interval(_interval)
-        {
-            startTime = std::chrono::system_clock::now();
-            progress = 0;
-            parent->AddRipple(this);
-            win = parent->getWindow();
-            nowText = "";
-            if (_interval < 0 /* so is -1 or bogus value*/) {
-                // assume it hasn't been set
-                looping = false;
-            }
-            else {
-                looping = true;
-            }
-        }
-
-        void eraseRipple() {
-            std::string textBlank;
-            for (int i = 0; i < text.length(); i++) {
-                textBlank += " ";
-            }
-            mvwprintw(win, y, x, textBlank.c_str());
-        }
-
-        void restartRipple() {
-            eraseRipple();
-            done = false;
-            progress = 0;
-            nowText = "";
-        }
-
-        void Update() {
-            auto now = std::chrono::system_clock::now();
-            std::chrono::duration<double> elapsed = now - startTime;
-            double elapsedSeconds = elapsed.count();
-
-            if (elapsedSeconds >= delay && !done) {
-                nowText += text[progress];
-                mvwprintw(win, y, x, nowText.c_str());
-                startTime = std::chrono::system_clock::now();
-                progress++;
-                if (progress == text.length()) {
-                    done = true;
-                    progress = 0;
-                }
-            }
-
-            if (looping && done) {
-                if (elapsedSeconds >= interval) {
-                    restartRipple();
-                }
-            }
-        }
-
-};
-
-std::vector<RipplePrint*> WindowClass::getRipples() {
-    return Ripples;
-}
-
-void WindowClass::setBorderColor(COLOR color) {
-    BorderColor = color;
-}
-
-void WindowClass::AddRipple(RipplePrint* rip) {
-    Ripples.push_back(rip);
-}
-
-WINDOW * WindowClass::getWindow() {
-    return win;
-}
-
-void WindowClass::InitWindow() {
-    if (win == nullptr) {
-        win = newwin(cur_h, cur_w, y, x); 
-
-        // running content in here instead of in the constructor so that that the callback can use fully updated values
-        // possible dynamic content positioning and rendering
-        if (content) {
-            content(this); // Function that adds all the objects to the be display to the window on initialization
-        }
-    }
-    else {
-        std::cerr << "Window has already been initialized" << std::endl;
-    }
-}
-
-void WindowClass::Normalize(int x_offset, int y_offset) {
-    x -= x_offset;
-    y -= y_offset;
-}
-
-void WindowClass::setSize(int height_stdunit, int width_stdunit) {
-    cur_h = height_stdunit * std_unit;
-    cur_w = width_stdunit * std_unit;
-    resize_window(win, cur_h, cur_w);
-}
-
-void WindowClass::setPos (int x_stdunit, int y_stdunit) {
-    int _x = x_stdunit * std_unit;
-    int _y = y_stdunit * std_unit;
-    if (mvwin(win, _y, _x) != ERR) {
-        x = _x;
-        y = _y;
-    }
-}
-
-int WindowClass::getHeight() {
-    return cur_h;
-}        
-
-int WindowClass::getRippleCount() {
-    return Ripples.size();
-}
-
-int WindowClass::getWidth() {
-    return cur_w;
-}
-
-int WindowClass::getX() {
-    return x;
-}
-
-int WindowClass::getY() {
-    return y;
-}
-
-bool WindowClass::IsShowing() {
-    return isShowing;
-}
-
-void WindowClass::IsShowing(bool set) {
-    isShowing = set;
-}
-
-bool WindowClass::IsSelectable() {
-    return isSelectable;
-}
-
-void WindowClass::IsSelectable (bool set) {
-    isSelectable = set;
-}
-
-void WindowClass::Update() {
-    if (isShowing) {
-        if (has_border) {
-            wattron(win, COLOR_PAIR(BorderColor));
-            box(win, 0, 0); // Draw border if needed
-            wattroff(win, COLOR_PAIR(BorderColor));
-        }
-        //auto now = std::chrono::system_clock::now();
-
-        for (auto rip : Ripples) {
-            rip->Update();
-        }
-
-        //std::chrono::duration<double> elapsed2 = now - std::chrono::system_clock::now();
-        //double elapsedSeconds2 = elapsed2.count();
-        //mvwprintw(win, 1, 1, "all ripples time: %f seconds / ripple count: %d", elapsed2, Ripples.size());
-
-        wrefresh(win); // Refresh the window to update display
-    }
-}
 
 void IntroMethod(WINDOW * win) {
     mvwprintw(win, 1, 1, "Welcome to the Game!");
